Fix up debugger attach handlers
First off, the error messages were never actually displayed to the user
because the RpcErrorException constructor takes three arguments.
Secondly, we do not support attaching to PowerShell Editor Services. It
sure looked like we did (because we had special logic for it) but once
attached, nothing worked. So it was half-baked. Now we throw an error if
the user is trying to do that. Thirdly, because of that half-baked
implementation, the process ID field was typed as a string (to support
"current" as a shortcut)but that caused a mess here and an error in the
VS Code client. Now it's just always an integer. (Same for the runspace
ID.) Fourthly, a big mess was cleaned up by refactoring using functions,
who'd have thought? Fifth and finally, superfluous version checking
around PowerShell <5.1 was removed (as those versions are no longer
supported whatsoever).